Fix import structure by moving api/ into src/ticker_converter/
- Move api/ directory into src/ticker_converter/api/ for proper package structure
- Update run_api.py to use relative imports instead of path manipulation
- Fix dependencies.py to use relative import for database module
- Update SQL file path resolution in dependencies.py for new structure
- Remove sys.path manipulation from run_api.py (no longer needed)
- Fix asyncpg import issue in mypy configuration

This resolves pylint and mypy import errors by ensuring all modules
are properly organized within the main package structure.
