+++ src/ticker_converter/api/models.py
